+++ CHANGES.rst
@@ -3,6 +3,7 @@ v0.1.0 (in development)
 
 - memory footprint reduced
 - ``len`` support
+- ``CTrie#height`` method added
 
 v0.0.1
 ------